logger start setup errorcontroller ingredients can't be deleted if assigned to pizzas

+++ src/main/java/de/telran/pizzaproject/controller/IngredientsController.java
@@ -31,8 +31,8 @@ public class IngredientsController {
 
     @PostMapping
     public String addIngredientsDetails(@ModelAttribute("ingredientToAdd") @Valid Ingredient ingredientToAdd,
-                                  BindingResult result,
-                                  RedirectAttributes attributes, Model model) {
+                                        BindingResult result,
+                                        RedirectAttributes attributes, Model model) {
         validator.validate(ingredientToAdd, result);
         if (result.hasErrors()) {
             model.addAttribute("ingredientToAdd", ingredientToAdd);
@@ -54,9 +54,9 @@ public class IngredientsController {
     @PatchMapping("/update/{id}")
     public String updateIngredientsDetails(
             @ModelAttribute("ingredientToUpdate") @Valid Ingredient ingredientToUpdate,
-                                     BindingResult result,
-                                     Model model,
-                                     RedirectAttributes attributes) {
+            BindingResult result,
+            Model model,
+            RedirectAttributes attributes) {
         validator.validate(ingredientToUpdate, result);
         if (result.hasErrors()) {
             model.addAttribute("ingredientToUpdateId", ingredientToUpdate.getId());
@@ -69,7 +69,11 @@ public class IngredientsController {
 
     @DeleteMapping("/delete")
     public String deleteIngredient(@RequestParam("ingredientId") Long ingredientId,
-                                     RedirectAttributes attributes) {
+                                   RedirectAttributes attributes) {
+        if (service.getIngredientById(ingredientId).getPizzas().size() > 0) {
+            attributes.addFlashAttribute("ingredientBlocked", ingredientId);
+            return "redirect:/ingredients";
+        }
         service.deleteIngredient(ingredientId);
         attributes.addFlashAttribute("deleted", ingredientId);
         return "redirect:/ingredients";
@@ -77,6 +81,6 @@ public class IngredientsController {
 
     @ModelAttribute
     public void addAttributes(Model model) {
-         model.addAttribute("ingredients", service.getAllIngredients());
+        model.addAttribute("ingredients", service.getAllIngredients());
     }
 }
